requirements.txt
astroid==2.5.6
click==8.0.3
cycler==0.10.0
Flask==2.0.2
imbalanced-learn==0.8.0
imblearn==0.0
install==1.3.4
isort==5.8.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib==3.4.2
mccabe==0.6.1
numpy==1.21.0
opencv-contrib-python==4.5.3.56
pandas==1.2.5
Pillow==8.3.0
pylint==2.8.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
six==1.16.0
threadpoolctl==2.1.0
toml==0.10.2
Werkzeug==2.0.2
wrapt==1.12.1
